requirements.txt
torch>=2.0.0
transformers>=4.30.0
pymongo>=4.0.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
seqeval>=1.2.0
tqdm>=4.60.0
numpy>=1.20.0
